Fix flaky legacy hashing test
Our legacy `HashTest.test_hash_funcs_acceptable_keys` is flaky. We create two instances of an unhashable type, and hash them with a custom `hash_func` that just uses the Python `id` builtin. 

`id` returns an object's unique identifier - two different objects will never share the same `id`. This particular hashing test seems to be failing with some regularity on Python 3.11, which is surprising: I would have expected the test to always fail on _all versions_ of Python (because we're hashing the IDs of two distinct objects).

This PR modifies the test to hash only a single instance of the given unhashable object; the test now passes on Python 3.11. (Why the test passed on earlier versions of Python remains a mystery!)