Fix issue #6310 (#6321)
check that cache directory exists before trying to list files there.

we try to remove all files in the cache directory that start with the function key, whether clear is called for self.persist
storage or not, to avoid leaving orphaned files in the cache directory. This leads to a bug, when cache directory is not exist, but we try to list all files there (to filter only those that belong to the corresponding function).

The solution is to check that directory exists before trying to filter corresponding files there.