Refactor Duplicated Export Names (#6660)
For st-lib, I was facing an eslint error where proto exports Button and we export Button through shared/button. As a result, we want to deduplicate the naming and I propose renaming shared/Button to BaseButton.

This happens for elements/Alert and shared/ColorPicker. This PR is only for the ones that st-lib needs but maybe we can add the rest of them into the tech debt db such as Widgets/Button and other Props.