Feature: st.toast (#6783)
Adds a new widget for notification toasts using st.toast - covers the use case for notifications that don't disrupt the flow/organization of the existing app. This is distinct from the existing st.warning, st.error, st.success & st.info, which are inline. 
NOTE: This PR also creates a new event rendering construct in an effort to distinguish events from persistent elements like widgets. The path is as follows:
- new RootContainer constant -> EVENT
- lib/streamlit/__init__.py -> toast mapped to event.toast vs. other methods on main
- AppView.tsx has a new EventContainer that uses existing renderBlock method that takes in event elements