Add ability to click enter on textInput, NumberInput, and TextArea when in st.form (#6911)
Extra Context
We are not going to be adding a <form> for now. Why?
When you add a <form>, it was submitting two clicks because inputs inside of form automatically find the next button and click it.
Because <input> inside of a <form> automatically find the next <button> and call onClick , this causes problems when there’s a st.number_input as there are buttons to do the - and the + functionality.

Example html to see it just clicks on the next button when clicking enter:

<!DOCTYPE html>
<html lang="en">
  <head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <link rel="stylesheet" href="src/style.css">
  </head>
  <body>
    
    <h1>Hi friend, try edit me!</h1>
    <form onsubmit="(event)=> event.preventDefault(); return false;">
      <label for="name">Enter username (upper and lowercase letters): </label>
      <input type="text" name="name" id="name" required pattern="[A-Za-z]+" />
      <button onclick="console.log('Hello !')">Submit</button>
      <button onclick="console.log("Hello 2!!")">Submit2</button>
    </form>


  </body>
</html>
Describe your changes
The changes mainly change WidgetStateManager.ts.

Instead of keeping track of the submitButtonCount, we keep a map of Array<SubmitButtonProtos>.
We map the formId to that Array<SubmitButtonProtos>.
Added the ability for any widget to submit form in util/utils.ts
Show Press Enter to submit form when we are in st.form
Modify st.chat_input but maybe that might not be necessary and I can revert those changes.
Demo: https://enter-submit-form.streamlit.app/#inside-a-st-form-where-clear-on-submit-false
GitHub Issue Link (if applicable)
#3790

Testing Plan
Explanation of why no additional tests are needed
Unit Tests (JS and/or Python)
added and changed corresponding Unit Tests
E2E Tests
I don't think these need any e2e tests as the unit tests should cover it.
Any manual testing needed?
None
Contribution License Agreement

By submitting this pull request you agree that all contributions to this project are made under the Apache 2.0 license.