Remove local e2e environment (#8087)
## Describe your changes

The local e2e test env is unused and mostly out-of-date. It's still
running on Python 3.7, but Streamlit no longer supports this. And it is
not really needed for the new Playwright setup. Therefore, this PR
removes the relevant artifacts form the repo.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.