Refactor plotly theme creation (#8110)
## Describe your changes

A small refactoring related to the creation of the plotly theme. Instead
of using a module for the plotly theme, we are capturing the creation in
a method to make it slightly better behave to enable lazy-loading for
plotly.

There aren't any other logical changes in here related to the Plotly
theme.

This refactor also applies a couple of other small refactorings related
to the imports.

Related to #6066

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.