Update `browser.serverPort` config option (#8139)
## Describe your changes
`browser.serverPort` was deprecated in PR #8118 but this PR reversed
that since an isolated use case for the option was identified. It is
used for the file upload endpoint if XSRF protection is activated.

## Testing Plan
None. This just reverts a deprecations.

---

**Contribution License Agreement**
By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.