Show a warning when server port `3000` is used (#8152)
## Describe your changes

Port 3000 is reserved for internal development. This PR adds a warning
to make it more obvious to users to use a different port.

Related to https://github.com/streamlit/streamlit/issues/8149

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.