Remove .ttf font files (#8171)
In #7157, we replaced the usage of the `ttf` font format with `woff2`, which offers
better file compression. We didn't delete the `ttf` files at the time just in case we
had to revert back due to some unanticipated reason, but now that it's been a few
months it's safe to remove the unused files.