Make Text Input Ids more stable for testing purposes (#8286)
## Describe your changes

The id for the widget label changes every rerun. This doesn't have a
problem in functionality, but it has a problem with our text where the
id changes in a `getByLabelText`

## Testing Plan

- Previous unit test will apply, and I added another such that a change
in the value does not change the htmlFor label.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.