Add linting to ensure lodash imports are correct (#8287)
## Describe your changes

We made a change on how we import lodash to avoid builds from pulling
the whole library. To fix this, I added an eslint plugin so that it
ensures the correct import happens.

## Testing Plan

- Initial tests should show the failure
- Tests after merging in develop (with the corrected changes) should
show it working.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.