Fix issue with not correctly waiting for connections (#8294)
## Describe your changes

[In this PR](https://github.com/streamlit/streamlit/pull/8068), we
accidentally removed the waiting for `has_connection.` This means that
an app without any connection is likely to just do a lot of necessary
loops, causing higher CPU usage compared to previous versions. This PR
adds the `has_connection` await.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.