Fullscreen Button Overflow Horizontally (#8279)
## Describe your changes

As of now, our full screen button goes outside of the boundary, which in
some viewports can cause overflowing widths to make the container
scrollable. We intend to make the full screen button experience not as
bad, but for now, the solution is to reduce the content area size when
we hit this threshold.

## GitHub Issue Link (if applicable)
Closes #6990 

## Testing Plan

- E2E Test to capture the look created from the bug.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.