Offset dates in vega if no timezone information is attached. (#8278)
## Describe your changes

When Streamlit apps work with dates/datetime, Vega can show the
incorrect date on the graphs. This is because datetimes are serialized
as milliseconds since epoch and are therefore assumed in UTC format.
Vega's engine takes in this timestamp and produces a Date that is set to
the local timezone, which can create some incorrect values. There's also
no way to determined from the serialized shape what timezone it is in,
so we need to offset the time for vega.

NOTE: We do this a lot in formatting for tables, but this is specific to
Vega's engine, so we isolate the change there.

## GitHub Issue Link (if applicable)
Closes #4342 

## Testing Plan

- JS Unit Tests are included and updated to test DATE, DATETIME, and
DATETIMETZ formats.
- Checking rendering of E2E Tests for the tooltip (like in the issue) is
really difficult. Manual testing of this is possible with PR Preview.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.