Fix Pillow type errors (#8412)
## Describe your changes
Access the bilinear option via the modern Resampling enum.
Indicate that raw bytes are allowed for AtomicImage, to resolve an
unreachable statement error. I chose to add this because the code and
tests indicate that it should be supported, but possibly it should not
be allowed and that code path should be removed instead.


## Testing Plan

Not a behavioral change, no new tests needed.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.