Update Codeowners to represent the team (#8590)
## Describe your changes

More people have joined the team, and an update to the CODEOWNERS file
is needed. We created a Release Team whose focus is to be a quality
control for these things. It is composed of senior engineering
leadership. We also added any individual who has implemented a core of
the section as an owner of the area.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.