Remove default value for title of dialog_decorator (#8729)
## Describe your changes

The default value for the dialog_decorator title `""` looks like it is
not-required. But truth is that we throw an error for `""` stating that
no empty title must be provided. Hence, we remove the default value.

## GitHub Issue Link (if applicable)

## Testing Plan

- Since the value `""` was never allowed, removing the default should
not have an impact on users. It is mainly to improve our docs.
---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.