Add workflow that only runs playwright tests of changed files (#8739)
## Describe your changes

This workflow only runs playwright tests that have been changed within a
pull request. This makes it a lot faster to get the relevant test
results and screenshots. This will also create trace files for all
failed tests that contain a lot more detailed information about the
failed test.

To test this workflow, I added a couple of reasonable changes to e2e
tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.