Change the `fragmentId` to be a required property in `WidgetStateManager` (#8533)
## Describe your changes

The `fragmentId` is currently an optional property for all the set state
methods in the `WidgetStateManager`. The reason here is that it
sometimes has a value and sometimes is just `undefined`. But this makes
it a bit too easy to forget to add it -> which can lead to features that
are not compatible with fragments.

This PR changes the property to required by changing from `fragmentId?:
string` to `fragmentId: string | undefined`. This also uncovered a
potentially existing issue with form submission & fragments, which will
be fixed via this PR, e.g.:


https://github.com/streamlit/streamlit/pull/8533/files#diff-b6b2b8d7411dc40d04d4dc0442572c4835e84a79844cd2e299a469b7b2ec1a1cR309

## Testing Plan

- Update tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Vincent Donato <vincent@streamlit.io>