Ensure SessionInfo is set before performing an action (#8779)
## Describe your changes

Reports of SessionInfo not being set have increased. Latest findings
suggest the ScriptFinished task is still being run despite the
NewSession request not running first. I suspect this may be due to #8599
filtering out messages and keeping the ScriptFinished but not keeping
the NewSession.

In _any_ case, this is an error we should not be surfacing to users.
This change guards the bad code, and sends a metric for an error
otherwise so that we can continue to track the issue and find a better
solution.

## GitHub Issue Link (if applicable)
closes https://github.com/streamlit/streamlit/issues/8321
closes https://github.com/streamlit/streamlit/issues/7549

## Testing Plan

- Updated JS Unit Tests to ensure no issues occur

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.