Remove check for new version of Streamlit (#8841)
## Describe your changes

We want to remove a specific request to PyPi as we did not believe it
was providing enough value. It's best to remove and in the event we want
to re-introduce it more thoughtfully, we can.

## GitHub Issue Link (if applicable)

closes https://github.com/streamlit/streamlit/issues/8453

## Testing Plan

- This is mostly a removal of code, so original tests should still pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.