Show fragment errors in fragment-path for main app runs (#8868)
## Describe your changes

When fragments are executed in context of a main app run, their code is
executed as a direct function call
[here](https://github.com/streamlit/streamlit/blob/develop/lib/streamlit/runtime/fragment.py#L194).
If the fragment body raises an exception, the try-except block from the
main script is used
[here](https://github.com/streamlit/streamlit/blob/develop/lib/streamlit/runtime/scriptrunner/script_runner.py#L593-L612).
This results in the fact that those exceptions are shown in the main
page's path, which becomes very visually apparent when using dialogs,
where the error is shown outside of the dialog.

The PR refactors the code execution from script_runner and calls the
wrapped code from `fragment.py`, so that the try-except wrapper is
always the same for a fragment, no matter whether it is executed in
context of a main app run or as a fragment re-run. Note that for a
fragment rerun, the body was already executed in the try-except wrapper
[here](https://github.com/streamlit/streamlit/blob/develop/lib/streamlit/runtime/scriptrunner/script_runner.py#L581).

If a `RerunException` happens in the fragment code during main app
execution (=not a rerun solely for the fragment), the `RerunException`
is raised again so that the ScriptRunner try-except can catch it and
trigger a rerun appropriately; which basically is the same behavior as
before the change.

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.