Add support for numpy 2.x (#8940)
## Describe your changes

Numpy 2.0 was released a couple of days ago. This PR adds support for
numpy 2.x.

Streamlit not supporting numpy 2.0 likely leads to the side-effect that
Streamlit 1.19 gets installed by the resolver if Streamlit & numpy are
unpinned. The reason:

We added major version pins in Streamlit >= 1.23. This can lead to the
interpretation (by the resolver) that Streamlit < 1.23 is compatible
with numpy 2.0, but >= 1.23 isn't. So maybe the resolver selects Numpy
2.0, and then resolves to a compatible Streamlit version which would be
< 1.23. But the version that actually gets selected is 1.19 because we
added a major version pin to altair <5 in 1.20. Thereby, the latest
Streamlit version that is semantically compatible with Numpy 2.x and
Altair 5.x is 1.19.

## GitHub Issue Link (if applicable)

- https://github.com/streamlit/streamlit/issues/8927
- https://github.com/streamlit/streamlit/issues/8917
- https://github.com/streamlit/streamlit/issues/8935
- https://github.com/streamlit/streamlit/issues/8936


## Testing Plan

- Not logic changed -> no tests are required. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.