Add script to update material icon names (#8996)
## Describe your changes

Add a script to update the material icon (symbols) names. In a second
step, we can also add a CI job that executes this automatically and
fails as soon as there are changes to indicate that the font can be
updated. This can be coupled with the `update_emoji` script as well.

## Testing Plan

- Not logical changes

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.