READ only headers and cookies (#8976)
## Describe your changes

Add a new `st.context` command with `headers` and `cookies` attributes. 
They are dictionary-like objects (for headers, case-insensitive, with
`get_all` method to get a list of headers in case the same header is
repeated in websocket request).

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python) Added
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.