Support theming in embed better (#9103)
## Describe your changes

This change resolves three challenges:

1. The "Auto Theme" respects embed params first before using a system
theme. We use the AutoTheme as an "undo" of any theme. We can perhaps
consider calling it a "Default Theme"
2. For post message, we will allow setting Light or Dark theme and not
make it interpreted as a custom theme. We do that by specifying the
theme name or a custom theme.
3. Theme specified embedded apps disables cached theming. Cached theming
essentially occurs when the developer defines a custom theme or when the
user specifies a theme preference in the settings dialog. Lots of
confusion from how the embedder wishes apps to look causes issues, so
it's best to not cache any theme when there's a preference from the
embedder. The tradeoff is that custom themes will not be cached for the
next rerun, but it feels like the right set of controls to ensure proper
functionality.

## Testing Plan

- JS Unit Tests are updated

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.