Fix ghost tabs issue in fragments (#9186)
## Describe your changes

Closes https://github.com/streamlit/streamlit/issues/9158

So far we seemed to have missed the case where we can have stale widgets
in the same fragment we are currently running which stem from a previous
run. This PR closes this gap by removing widget nodes that belong to the
current fragment but an old run.
It looks like the reason why this is happening for tabs (and probably
other containers) and not for other elements is that in for a tab delta,
the `scriptRunId` is updated to the current run, so the existing pruning
logic does not work; the new approach also considers the child's
`scriptRunId`!

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
  - Extended unit test to simulate this issue
- E2E Tests
  - Add e2e test with example app from #9158
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.