Fix issue with empty Azure OpenAI chunk (#9234)
## Describe your changes

Don't error out on message chunks in `st.write_stream` if the delta
property is None. Based on
https://github.com/streamlit/streamlit/issues/9227, this can sometimes
happen in the Azure OpenAI client.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/9227

## Testing Plan

- We are not yet integration testing usage of the Azure OpenAI client
since this would require an active connection.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.