Fix mypy errors (#9254)
## Describe your changes

Our mypy-checks have started to fail in the merge gates (see
[here](https://github.com/streamlit/streamlit/actions/runs/10347072909/job/28639963961)),
probably because altair was updated with better typing. This PR updates
our max version of mypy and fixes some of the reported errors with
`mypy==1.10`

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.