When clearing forward queue add a stopped for rerun message (#9187)
## Describe your changes
At the moment, some delta messages may not be sent to the frontend but
the script_finished_successfully message may still be sent to the front
end, at which point the front end will remove stale elements.

From the perspective of the frontend there is no difference between a
scriptRunner finishing execution but some of the later messages being
dropped and the scriptRunner being stopped prematurely for a rerun.

In the case of the issue #9163, the number input element is considered
stale, so removed. It is then added back by the rerun, but has lost its
state, so the value is nulled out.

This PR instead modifies the preserved script_finished messages by
setting FINISHED_EARLY_FOR_RERUN as the value when the browser queue was
not flushed before a new script run started.

## GitHub Issue Link (if applicable)

Closes https://github.com/streamlit/streamlit/issues/9163

## Testing Plan

The bug this aims to fix is related to concurrency and seems quite
difficult to test in an automated fashion. The bug cannot be reliably
reproduced so e2e testing doesnt seem too useful either.

I have verified these changes resolve the issue shared above. 

Updated the existing tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

Co-authored-by: Dan Allford <daniel.allford@bp.com>