Use overloads to make the return types of st.feedback more precise. (with type tests) (#9216)
## Describe your changes

Inspired by the adventures of @Asaurus1 in issue #9048, I thought it
would be nice for the return type of st.feedback to be narrowed so that
my typechecker knows if the function returns 0|1|None or 0|1|2|3|4|None.
So, I used `@overload` and `Literal` to achieve this.

## Testing Plan

@Asaurus1, in his pr, is introducing a folder of test code that mypy now
gets run over to verify that the return types are correct. This seemed
like a good idea to me, so I used the same path and concept.