Fix issue with re-adding deleted row in data editor with same index (#8864)
## Describe your changes

Fixes an issue with re-adding a deleted row in the data editor with the
same index. The fix moves the handling of added rows below the handling
of row deletions. This fix also requires an update of the minimum
supported Pandas version to 1.4, since < 1.4 has a bug where the index
is changed from `RangeIndex` to `Int64Index` if a row is dropped.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/8854

## Testing Plan

- Added unit test

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <103002573+sfc-gh-lmasuch@users.noreply.github.com>