Migrate iframe_resizer e2e test from Cypress to Playwright (#9433)
## Describe your changes

Migrate the existing test to Playwright.

Update the convenience function for loading a Streamlit app in an iframe
to take additional arguments to make it configurable. In this case,
allow
- injecting the iframe resizer script
- add query params to the source
- add id to iframe html element

Also, exclude our `e2e_playwright/test_assets/` folder from the license
check.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
  - Migrate the existing e2e test
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.