Migrate hostframe test from Cypress to Playwright (#9436)
## Describe your changes

Migrate the test from Cypress to Playwright. This is the last test to be migrated.
Also, the new version of the app uses multi-page apps v2 instead of the
old one!

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
  - Migrate existing e2e test
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.