Fix accidentally watching parent folder for changes (#9453)
## Describe your changes

Streamlit accidentally watches the parent directory of the directory it
is supposed to watch. This can get quite problematic since the parent
directory can contain lots of files. This PR fixes it by correctly
resolving the path.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/7467

## Testing Plan

- Added unit test to make sure paths are correctly resolved.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.