Fix dataframe maximum update depth error  (#9490)
## Describe your changes

Fixes an issue with the dataframe component that causes random crashes:


![image](https://github.com/user-attachments/assets/da32e051-95ac-4562-aee6-bdb82a2fbd27)

The crash can happen on low-resolution monitors if dataframes are
rendered with `use_container_width=True` are used in some of our
containers (popover, tabs, sidebar, columns). It appears that our width
calculation in the Block component causes this. If the width can't be
determined yet by the resize observer, it can be -1 for some time. This
seems to be a trigger for the maximum update depth. Unfortunately, I
don't have a validated explanation of why it actually leads to the error
:(

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/7949

## Testing Plan

- Add e2e test to discover similar breakage. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.