make useWidgetManagerElementState respect form clearing (#9591)
## Describe your changes

This PR addresses one gap in functionality between
`useWidgetManagerElementState` and `useBasicWidgetState` -- handling
when a form gets cleared. The next PR will then attempt to merge these
two hooks together so that we don't have similar but different hooks for
essentially the same thing.

## Testing Plan

- Unit Tests (JS and/or Python)
  - added unit tests


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.