Fix clear cached class method regression (#9642)
## Describe your changes

https://github.com/streamlit/streamlit/pull/9101 caused a small
regression with clearing the full cache of a cached class method. This
PR fixes it by only passing the self instance if the clear method
actually has args or kwargs defined.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/9633

## Testing Plan

- Added unit tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.