Improve hover tooltips for `st.line_chart` (#9674)
## Describe your changes

Improves the support for hover tooltips on line charts by using a
layered chart. This will allow us to show markers and tooltips for the
nearest object on the line chart.

- [Demo](https://better-line-tooltips.streamlit.app/)

## Testing Plan

- Updated unit tests
- Added e2e test with snapshots

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.