Make virtual dropdown responsive to font size (#9842)
## Describe your changes

Makes the virtual dropdown (used by selectbox and multiselect) size
responsive to the root font size by using `rem` instead of using `px`
size. This is required for the upcoming changes for advanced theming.

## Testing Plan

- No logical or visual changes.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.