Fix issue with missing test-results in the e2e playwright workflow (#9876)
## Describe your changes

There is an issue sometimes that the playwright output directory gets
deleted in between the pytest session. It seems to be some kind of issue
with the combination of pytest-playwright, xdist and
pytest-rerunfailures. To prevent the issue, we deactivate the
`delete_output_dir` behavior if reruns and xdist is activated. The
directory gets deleted anyways in our makefile before running the tests.

The PR also contains a couple other small drive-by improvements to make
our tests more robust.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.