Add async generator support to `st.write_stream` (#8724)
## Describe your changes

Implements support for async generators when used as input for
`st.write` and `st.write_stream`. This is not adding real async support,
but instead just converting async generators into a sync generators.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/8161

## Testing Plan

- Added e2e & unit tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.