Preserve target='_blank' for html elements (#9994)
## Describe your changes

Closes https://github.com/streamlit/streamlit/issues/9972

DOMPurify removes the `target` attribute on element nodes. Based on
this Google doc
article(https://developer.chrome.com/docs/lighthouse/best-practices/external-anchors-use-rel-noopener)
and this conversation in the DOMPurify
repository(https://github.com/cure53/DOMPurify/issues/317), it is
secure to keep "target=_blank" when at the same time "rel=noopener
noreferrer" is set. So we use a DOMPurify hook to preserve the attribute
if the user set it

## GitHub Issue Link (if applicable)

https://github.com/streamlit/streamlit/issues/9972

## Testing Plan

- Added unit tests to ensure the sanitization behavior

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.