Make st.write use st.html when _repr_html_ is present. (#9910)
Make st.write use st.html when _repr_html_ is present.

## Describe your changes

1. When `_repr_html_()` is present in an object, `st.write` now displays
the HTML using `st.html()` rather than `st.markdown()` (assuming
`unsafe_allow_html=True`).

1. When `unsafe_allow_html=False` AND `_repr_html_()` did not contain
HTML tags we used to send its output to `st.markdown()`. This behavior
is now removed. Instead, the output is handled with `st.help()` like any
other object.

1. Removes `_repr_html_()` from `*Connection` objects. This wasn't
really buying us anything, and was just weird. Also, this was the reason
for adding behavior #2 in the first place.

1. This is totally unrelated to the above, but is so small I didn't want
to create a separate PR for it (though, LMK if you want me to!): Remove
an old check for an environment variable that Streamlit's Atom plugin
from 2018-19 used to depend on. That plugin no longer exists, so we can
safely remove this.


## GitHub Issue Link (if applicable)
n/a

## Testing Plan

- ~~Explanation of why no additional tests are needed~~
- Unit Tests (JS and/or Python): ✅ Included!
- E2E Tests: 🟢  No new tests added, but this is already well-tested.
- Any manual testing needed? 👎 No need.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.