Fix CI issues with ubuntu 24 update (#10116)
## Describe your changes

The `ubuntu-latest` version got updated to ubuntu 24.x which breaks two
of our workflows:
- `pre-preview`: fixed by installing awscli via pip.
- `min-depts`: fixed by enforcing Ubuntu 22.04. The alternative solution
would be to use a higher Pillow pin. But it kind of makes sense to run
our min dependency tests also with older versions of Ubuntu.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.