Make st.expander summary respect isStale. (#10085)
## Describe your changes

Today, when an app reruns, widgets turn slightly transparent to denote
`isStale`. However, `st.expander`'s summary doesn't follow this behavior
(i.e. the header that always shows, even when the expander is
collapsed).

This PR makes `st.expander` respect `isStale` like everything else.

### Before


https://github.com/user-attachments/assets/0cb5291c-4795-4993-9556-d0a3629ea1ce


### After


https://github.com/user-attachments/assets/d82ad1ae-e2c2-42b5-977e-e69690d32642


## GitHub Issue Link (if applicable)

None

## Testing Plan

- Explanation of why no additional tests are needed:
This behavior can only be properly tested with screenshot tests, but
since the `isStale` behavior depends on timing and animations, any
screenshot test is going to be flaky. The only way to reduce flakiness
would be to make the test longer, and in fact I did exactly that with a
10s-long `time.sleep` a few moments ago. But I decided not to commit the
test because I don't think this test case is worth the cost of slowing
down our test suite by 10s.

- Unit Tests (JS and/or Python): No
- E2E Tests: No
- Any manual testing needed? No

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.