Refactor arrow type utils to centralize type handling (#10050)
## Describe your changes

Refactors the arrow type util to include more utility methods to work
with arrow types. This is intended to move all logic on how to handle
arrow types to the quiver module. This is to prepare for some of the
foundational dataframe improvements.

This is only the first step for refactoring this aspect, there will be
subsequent PRs to centralize the type handling even more.

## Testing Plan

- Added unit tests for all new exported functions.
- Existing e2e & unit tests should cover this well that it doesn't apply
any end-user-facing changes.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.