Refactor vega lite chart component to a React functional component (#9959)
## Describe your changes

Converted this to a Function component and simplified it significantly
(I believe). I divided it into three hooks.

1. One that preprocess the element and provides more stable references.
2. One that provides the lifecycle methods of create, update, and
finalize view.
3. One that handles interactivity with forms and whatnot (leveraged in
the `useVegaEmbed`

In order to detect data changes, I generate a hash of the data.

## Testing Plan

- Current tests should pass. Will add more for hooks to be more
thorough, but I wanted to see if the structure works for everyone.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.