Show elapsed time on `st.spinner` (#10072)
## Describe your changes

Adds a parameter `show_time` to `st.spinner`. If set to True, the
spinner shows the elapsed time in a small gray font next to the spinner
text. Defaults to False (i.e. same behavior as today).

![CleanShot 2024-12-24 at 01 11
31](https://github.com/user-attachments/assets/e947ccaf-f5e3-413c-ba4a-50e9aa7547f6)

If we're going towards minutes and hours, the timer text will be
formatted nicely:

![CleanShot 2024-12-24 at 01 12
40@2x](https://github.com/user-attachments/assets/bdca9793-d054-41b0-bab7-684523150b3e)

## GitHub Issue Link (if applicable)

Closes #6805

## Testing Plan

Added unit tests for JS and Python and one simple e2e test. I tried to
make it not flaky but since that checks what's displayed after one
second, we need to see in practice whether it's flaky and potentially
adapt.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>