Remove shortcut to data checking and substitute with a rough hash (#10125)
## Describe your changes

We have a shortcut included to guess if the data is an extension of the
original data. This has caused issues in ensuring the data works as
expected. We want to make sure we only do this if the data is newly
added.

## GitHub Issue Link (if applicable)
Closes #6689

## Testing Plan

- Manual testing for the following:
   - add rows triggers an insert and not a new data
   - The same data does not update the data.
---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.