Make widget dropdowns look more consistent + convert `BaseColorPicker` to functional component (#10083)
## Describe your changes

Many of our widgets have dropdowns but they look slightly different.
This PR makes them look more consistent.

While I was at it, I also converted `BaseColorPicker` to a functional
component because I couldn't figure out how to fix some issues with
theming in the class-based component. The only things that changed in
the color picker are the injection of custom styles and the placement of
the popover, so I think doing the conversion at the same time shouldn't
be too confusing. But I can separate both changes if it's easier to
review.

| Before | After |
|----------|----------|
| ![CleanShot 2024-12-26 at 22 19
26](https://github.com/user-attachments/assets/8c7a4cc5-f169-4fbc-8256-8e9e8a39b5e8)
| ![CleanShot 2024-12-26 at 22 20
04](https://github.com/user-attachments/assets/fbb48ab0-8296-4632-9885-e9ec55b0402d)
|
| ![CleanShot 2024-12-26 at 22 20
37](https://github.com/user-attachments/assets/eb2d6172-d08c-4d54-9533-029197946c90)
| ![CleanShot 2024-12-26 at 22 20
57](https://github.com/user-attachments/assets/38fed40e-6582-4b52-a4e8-fc366b8cce92)
|
| ![CleanShot 2024-12-26 at 22 21
45](https://github.com/user-attachments/assets/f4f97485-c4c5-4137-9b72-3e8d2651aa0d)
| ![CleanShot 2024-12-26 at 22 22
21](https://github.com/user-attachments/assets/0b97eba6-9e1a-49f5-b4b6-122c01fff086)
|
| ![CleanShot 2024-12-26 at 22 24
24](https://github.com/user-attachments/assets/b3b3a6bd-8b27-439f-8cb9-44673196b3b5)
| ![CleanShot 2024-12-26 at 22 24
08](https://github.com/user-attachments/assets/9e390b46-7e30-464c-8320-d2c0d291835d)
|
| ![CleanShot 2024-12-26 at 22 24
55](https://github.com/user-attachments/assets/7c5212c6-7454-4554-9b1a-a76e33a321de)
| ![CleanShot 2025-01-13 at 01 10
48](https://github.com/user-attachments/assets/4679bd63-1965-4901-ae81-f1212069d277)
|

Note that the date input has a red border and is in the foreground when
focused (just like the other components), my screenshot tool somehow
removes that, though.


## GitHub Issue Link (if applicable)

Closes #8705 

## Testing Plan

- Updated snapshot tests for date input.
- I added a snapshot test for the dropdown of color picker since that
has a few customizations now.
- I realized we don't really have snapshot tests for the dropdown
elements of the other widgets. One problem is though that the snapshots
test only capture the content of the dropdown, not the shadow and
positioning around it. *Ideally*, we'd have some way to take a snapshot
with some additional padding around it but I'm not sure if it's worth
introducing that.
- Don't think it makes sense to have unit tests for such small styling
things.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Johannes Rieke <johannes.rieke@snowflake.com>