Make MPAv2 work with pure Python and with ProcessPoolExecutor (#10163)
## Describe your changes

Today, when `st.navigation` is present in an app, any [code using
`ProcessPoolExecutor`](https://discuss.streamlit.io/t/make-apps-faster-by-moving-heavy-computation-to-a-separate-process/68541#method-2-use-a-processpoolexecutor-5)
will throw an exception. Turns out this is because `ProcessPoolExecutor`
spins up a separate pure-Python process, and there's a bug that causes
`st.navigation` not to work in pure-Python mode (it requires a
ScriptRunContext).

This wasn't caught before because our pure-Python tests (`make
bare-execution-tests`) weren't actually running for scripts in
_subfolders_ of the main test folder.

...and when you run those previously-untested tests you discover
`st.page_link` _also_ doesn't work in pure-Python mode!

So this PR addresses all of the above:
1. Makes `st.navigation` and `st.page_link` work in pure-Python mode --
and therefore also when running inside a `ProcessPoolExecutor`
2. Makes `bare-execution-tests` test MPAv1 and MPAv2 code

## GitHub Issue Link (if applicable)

n/a

## Testing Plan

- ~~Explanation of why no additional tests are needed~~
- Unit Tests (JS and/or Python): ❌ 
- E2E Tests: ✅ 
- Any manual testing needed? ❌ 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.