Fix the chat input border radius (#10225)
## Describe your changes

The radius of the chat input border (only used in SiS) is currently
broken:


![image](https://github.com/user-attachments/assets/bcf92c25-0abb-4c30-9eb7-454ca6d6bb4d)

## Testing Plan

- We are not yet testing the advanced theming option of SIS. This is
something planned as follow up of the advanced theming project, but will
need a bit more effort to get proper tests set-up.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.