Use fixed px instead of rem for `contentMaxWidth` (#10236)
## Describe your changes

We don't want to adjust the content max width based on the root font
size, therefore, we are changing this setting to `px` instead of `rem`.
This allows us to fill the same screen estate regardless of the root
font size -> which allows more compact apps by using a small font size.

## Testing Plan

- Update hostframe snapshots

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.