Add e2e test for custom theming (#10248)
## Describe your changes

Adds an e2e test for our existing custom theming options (previously
untested). This test contains a mix of elements to as much of
interesting styling variations covered with the minimal number of
elements.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.