Apply top margin in top-aligned columns to first element instead of last (#10268)
## Describe your changes

We apply an additional margin for checkboxes & toggles to simplify
vertical alignment. However, for the margin was accidentally applied to
the last checkbox element instead of the first one for top alignment.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/10265

## Testing Plan

- Improved e2e test.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.