[Fix] Remove static app e2e test (temporarily) (#10286)
This PR removes the `static_app` playwright e2e/snapshot test - the static app conversion template is still being adjusted, and may cause breakage to this test. To avoid issues as we move into a release cycle, we are opting to remove the tests until the template is stabilized. In the meantime, there are still unit tests to cover the functionality.