Use compact file uploader based on element width (#10272)
## Describe your changes

Uses the element width instead of media breakpoints to determine if the
compact file uploader should be used. The previous version with media
breakpoints did not work well when file uploader is used in columns.

Before:

![image](https://github.com/user-attachments/assets/3639e5e6-a9ef-4363-a08e-ccc7a240ef3a)

After:

![image](https://github.com/user-attachments/assets/3b03235d-dc57-4127-8644-8841a77b8a9d)

## Testing Plan

- Added e2e test

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.