Activate venv before pip install (#10334)
## Describe your changes

Our Github actions fail because our venv is no longer being activated on
its own? As a result, we need to make sure we activate the venv each
time. I mapped out where we did that and source activate it to make sure
it works.

## Testing Plan

- CI should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.