Extract connection package (#10309)
## Describe your changes

A meaty PR. We had a lot of fun with small PRs to help decouple
packages, but at some point, we need to [draw the rest of the
owl](https://www.reddit.com/r/funny/comments/eccj2/how_to_draw_an_owl/)
(in kind terms 😄 ).

This change breaks the package responsible for the connection package.
Some scripts in version updates and pre-commit etc.

Several notes:

- Lib still requires connection in two places:
   - LibContext - Unsure where this should eventually go yet.
- Host Communication - we will likely create a host comm package (or
two, one for the host, and one for the client).
- I duplicated some functions 😲 . Most of them are incredibly tiny. I
think a follow up would be to have a utils library, and do it like
lodash of sorts (`import dateFunction from "@streamlit/utils/date"`,
`import { isNullOrUndefined, notNullOrUndefined } from
"@streamlit/utils/types"`, etc.)
- `yarn start` and building the app requires dependent libraries to be
built. connection is the first package that actually needs to be built
prior. As such, I modified the commands to build relevant packages as
necessary (some of them may duplicate effort, like protobuf generation),
but all those are relatively short and easy to do.
- Since it's the only one as of now, I did not try to make
auto-reloading work for all. IF you plan on watching that directory,
you'll need to run `yarn run buildWatch` in the connection package.
- I created a convenience `make frontend-dev` that ensures things build
then start. It also makes sure a sourcemap is generated

## Testing Plan

- Tests should pass.
- Verified manually that:
   - ✅ Hot reloading works with `yarn run buildWatch`
- For this, initial rebuild is about 2 seconds with rebuild being ~1
second. There's a small delay, but it does not seem bad for me.
      
Let me know if you want any other data to help understand the new
process.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.