Fix number input rounding and border (#10344)
## Describe your changes

Applies two small fixes to number input so that it correctly handles the
roundness from theme & that it uses the correct size if the border is
not activated. Currently, instead of showing the background color as a
border if not configured, it should use the secondary background color
instead so that the border looks like it is part of the input field.

## Testing Plan

- Added number input to theming e2e tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.