Fix the usage of dataframe column menu in dialog (#10359)
## Describe your changes

The dataframe  column menu is broken when used in dialog or popover:
 

![image](https://github.com/user-attachments/assets/7f430dfb-a7c0-4b84-9dfd-72f3b4c4bcc9)

The reason is that the `position: fixed` doesn't work in this context
since the dialog uses a transform -> which influences the position
fixed. To fix this, we are putting the column menu inside the dataframe
portal div that is also used for the cell overlays.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/10357

## Testing Plan

- Added e2e test

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.