add transparent file upload drop target to chat input (#10395)
## Describe your changes

We're making some changes to the chat input widget, primarily for the
file upload drag-n-drop behavior.

### What's new
* We opt'ed for showing the drag-n-drop area with the same dimension as
the chat input box, whether it's in single or multi-line mode. This
replaced the fixed sized drag-n-drop area with full-rounded corners
introduced in #9491.
* To help users drop files more easily when the chat input is short, we
added a transparent layer of larger drop target. This target has a
min-height of 90px, same as the fixed sized one before. If the chat
input box grows taller than that, the drop target simply follows its
dimension.


### Before
See https://github.com/streamlit/streamlit/pull/9491

### After


https://github.com/user-attachments/assets/71c4aad2-a4a5-44e3-9ea7-fe5a22614f85

### Debug mode (showing drop target borders)


https://github.com/user-attachments/assets/a03db11b-d037-4a82-b570-d5fbae4d8279


### Also fixed/improved in this PR
* Removed the reserved space for uploaded file previews above the chat
input box. This'd result in the content above the chat input box be
pushed up (or down) between having at least 1 uploaded file (or empty.)
But we figured that this is acceptable.
* Reworked the chat input box height to be 40px, our standard element
height. (It had become 42px due to borders being moved to the container
in #9491.) But since we still need to keep the borders on the container,
we introduced negative margins to children that need to aligned to top
or end within the container.
* Improved the drag-n-drop label content to distinguish between single
and multi-file modes.



## GitHub Issue Link (if applicable)

#7409 

## Testing Plan

- Unit Tests (JS and/or Python)
- E2E Tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.