Fix console error when using rem in svg width & height attributes (#10443)
## Describe your changes

react-feather icons render as svgs and the unit for width and height should be px instead of rem.

This PR transforms the `rem` to `px` in order to get rid of the error
warning.

## GitHub Issue Link (if applicable)

## Testing Plan

- No visual changes.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.