Update "click immediately with onBlur" test to reduce flakiness (#10450)
## Describe your changes

We have observed that doing the `st.write` in the callback _can_ lead to
flakiness for the following reason:
- the textArea's `onBlur` event triggers a rerun message and a rerun
- the rerun changes the number of rendered elements, because the rerun
triggered by the `textArea` does not include the button's callback
`write`
- the `click` event on the button is not dispatched by the browser,
because the click target changed

Now, the callback updates the session state and the `st.write` happens
outside of it, so that the number of elements per rerun is always the
same.
This still tests the original intention based on
https://github.com/streamlit/streamlit/issues/10007, because the
reported issue used a `print` statement and the issue was that the
trigger event happened before the text area rerun trigger.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
  - Updated e2e test to not be flaky
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.