Fix deprecation with `fillna` pandas method (#10484)
## Describe your changes

The way we use `fillna` is deprecated resulting in the following
deprecation warning:
Downcasting object dtype arrays on .fillna, .ffill, .bfill is deprecated
and will change in a future version. Call
result.infer_objects(copy=False) instead. To opt-in to the future
behavior, set pd.set_option('future.no_silent_downcasting', True)

This PR replaces the logic with a simple replace which should be
equivalent and probably also a bit more efficient.

## Testing Plan

- No logical changes, existing tests should cover this well.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.