Using LOG instead of log (#10492)
## Describe your changes

Part of some ongoing work to fix all violations of [react-refresh
rule](https://github.com/ArnaudBarre/eslint-plugin-react-refresh) which
will be followed by adding this eslint rule.

In some places we used `log` instead of `LOG` for the const. This
updates it to `LOG` so that it won't trigger the linting rule. Not all
are in `.tsx` files but I have just updated it everywhere for
consistency.

## GitHub Issue Link (if applicable)

## Testing Plan

Just renaming. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.