Change default for constraints to be turned off (#10512)
## Describe your changes

During releases, we hit constraints issues that require us to
cherry-pick updates that feel unnecessary. This change will disable them
for release (and most other operations). The decision overall is that we
do not need this for our CI, but we want something for users to rely on
for installing Streamlit reliably.

## Testing Plan

- Tests here should pass (and be well-formed)
- Tests verify env variable is not set

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.