Default to MPAv2 with changes and adjust tests we accept as graceful degradation (#10276)
## Describe your changes

This change switches the Default strategy on both the PagesManager and
AppNavigation. We discussed cases where this may produce breaking
behavior, but we would call these graceful degradation.

This effectively makes the V1 strategy dead code. A follow up PR will be
done to remove the dead code and refactor so it's a little bit cleaner.

## Testing Plan

- Tests should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: lukasmasuch <lukas.masuch@gmail.com>