Fix `font` theme config not being passed to custom components (#10661)
## Describe your changes

We deprecated the enum-based `font` property in our internal theme in
1.43. This also caused the `font` property to not be passed to custom
components that rely on `font` being filled. This PR fixes the
regression by passing in the new `bodyFont` property to the `font`
property so that it stays backwards compatible to existing custom
components.

## Github Issues

- Closes https://github.com/streamlit/streamlit/issues/10660

## Testing Plan

- Update unit tests.
- Manually verified the fix with the ant component.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.