Remove Deprecated Code Function (#10664)
## Describe your changes

In [this PR](https://github.com/streamlit/streamlit/pull/6042), we added
a new `def code` function in `code.py` which replaces the one in
`markdown.py`.

Since `CodeMixin` comes first in the definition of `DeltaGenerator`
before `MarkdownMixin`, the `CodeMixin` version of the function takes
precedence due to python's MRO algorithm and we can safely remove this
function.

## GitHub Issue Link (if applicable)

## Testing Plan

Covered by existing tests. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.