Use string path for page `__file__` (#10784)
## Describe your changes

When running the tests locally via `make pytest`, I get this error:
`AttributeError: 'PosixPath' object has no attribute 'endswith'` which
is related to us setting the page module path to a `Path` object instead
of a string path. Not sure why it works in our CI, but the `__file__`
property is expected to be a string path in Python.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.