[fix] Stabilize pydeck-select e2e tests (#10755)
## Describe your changes

- I noticed that the pydeck select tests were flaking a lot in CI. These
failures did not repro locally.
- I built a docker image and ran the tests locally and was able to
reproduce failures when the parallelization was high enough.
  - Note that CI runs with 32 Playwright workers!
- This made it clear to me that resource contention was at play here,
especially given that DeckGL is graphic intensive
- Previously `st_pydeck_chart_select.py` had 5 instances of DeckGL
rendering at a time. This PR changes it so that it only renders 1
instance of DeckGL at a time.
- After making this change, I have not been able to repro it locally in
my docker image, and I have yet to see these tests fail or re-run in CI


## GitHub Issue Link (if applicable)

## Testing Plan

- ✅ E2E tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.