Log script compilation error via error log level (#10826)
## Describe your changes

Currently, we only log script compilation errors via the `debug` log
level. This PR changes it to log it as an `exception` / `error` log
level instead. This simplifies usage in AI dev tools that access the
logging output.

## Testing Plan

- Added unit test.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.