Set text `color` for StyledSidebar (#10836)
## Describe your changes

Set `color` for `StyledSidebar` with `theme.colors.bodyText`, same as
[`StyledApp`](https://github.com/streamlit/streamlit/blob/develop/frontend/app/src/styled-components.ts#L27)


#### Context
We found an issue with multiple text components, in particular in
Markdowns, not respecting the

```toml
[theme.sidebar]

textColor = "red"

``` 
config in the sidebar. (sidebar theme was enabled in [this
PR](https://github.com/streamlit/streamlit/pull/10772).)
This was due to Markdown colors were inheriting from their parents and
never getting set by the theme. As a result, they were using the `color`
in `StyledApp`. Thus, in this PR we're overriding the `color` for
sidebar with the theme provided.

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.