Fix issue with background color handling in sidebar theming (#10854)
## Describe your changes

There was an issue with the current sidebar theme handling if
`backgroundColor` or `secondaryBackgroundColor` was configured to be an
empty string. This PR fixes this by correctly falling back to the main
theme colors (from emotion) on empty string.

## Testing Plan

- Added unit tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.