Fix boolean theme configs getting skipped (#10835)
## Describe your changes

Updated the condition in `_populate_theme_msg` to check if all theme
options values are `None` instead of checking if any values are present.
See my comment on how to repro and the root cause analysis.

## Testing Plan

- pytest


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.