Change calc_md5_with_blocking_retries to also except PermissionErrors (#10868)
## Describe your changes
Some IDEs and code editors may apply a lock when writing to files.
If auto-rerun is activated, the watcher triggers an acces to the file as
soon as it detects a change.
This leads to a PermissionError when streamlit tries to access that file
while the lock is still applied to it.
Adding the PermissionError to the excepted exception classes resolves
this problem, as the lock will usually be released before the function
reaches the maximum amount of execution attempts.

This PR essentially implements what was proposed in #4486.

## GitHub Issue Link (if applicable)
#4486 

## Testing Plan
Although there is no actual change in functionality, the affected
function now excepts a second exception class.
A corresponding unit test case might be needed.

---

**Contribution License Agreement**

By submitting this pull request I agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Karen Javadyan <kajarenc@gmail.com>