[refactor] Add enforce-memo eslint rule (#10871)
## Describe your changes

- Per a work item in our tech debt DB, which is aligned with the overall
performance workstream, this adds an `enforce-memo` eslint rule
- This eslint rule enforces that exported React components utilizes
`memo`
- The eslint rule also has an automatic fixer built in, making the
integration into IDEs much nicer and enabling `--fix` on the command
line
- Note: One of the intentional trade-offs is that we only check for
exported React components to have `memo` as a first pass. If we think
it's worthwhile to have _every_ React component in a given file utilize
`memo`, we can adjust the rule
- We only apply it to certain directories
- Note: This is another intentional trade-off for the first pass of this
PR. If we want to expand the directories, it is trivial to do so by
adjusting the `.eslintrc` file and running the eslint autofix

## GitHub Issue Link (if applicable)

## Testing Plan

- ✅ Adds unit tests for the rule definition

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.