Preserve selectbox frontend state when clicking away and don't rerun until selecting option (#10891)
## Describe your changes

This PR fixes an issue in `st.selectbox` where when you type and e.g.
remove some characters and then click outside of the selectbox, the
backend value will stay the same but the frontend selectbox is empty.
Also, it does not trigger a rerun until a value is selected.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.