Keep SessionInfo active after initialization, but include connection information (#9886)
## Describe your changes

We hit race conditions where the session info is cleared (on disconnect)
and reconnected before a page profile message is sent. We've hit
multiple issues with this, so the safest option is to not completely
delete the sessionInfo. In the event of the page profile, the message is
sent with the latest information it knows. We save the connection state
in the session info and update it to be disconnected on disconnect. It
will reset once a new NewSession message arrives.

## GitHub Issue Link (if applicable)
closes #9767

## Testing Plan

- Unit tests are updated

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.