[theming] Unify `st.help` border colors (#10916)
## Describe your changes

The borders in `st.help` between the gray bg header and the white bg
body use a lighter border color. Changing them to use our normal border
color to simplify the visual of the whole element.

## Testing Plan

- E2E Tests


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.