Don't disable element when empty options but accept_new_options is true (#10958)
## Describe your changes

Follow-up of https://github.com/streamlit/streamlit/pull/10788

When the initial `options` list is empty, we have so far disabled the
`selectbox` and `multiselect` elements. Though when `accept_new_options`
is `true`, the user should be able to add new options anyways.

## GitHub Issue Link (if applicable)

Closes https://github.com/streamlit/streamlit/issues/10948

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- Added unit tests to ensure that elements are not disabled and
placeholder has specific value when the passed `options` are empty but
`acceptNewOptions` is `true`
- E2E Tests
- Added e2e tests for `st_selectbox` and `st_multiselect` to ensure that
a user can type new options when `accept_new_options` is `False` even
when the initial `options` list is empty.
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.