Optimize `st.write` usage with single string arg (#10985)
## Describe your changes

If `st.write` is called with a single string arg, we directly pass this
to `st.markdown` instead of applying our buffer logic + `st.empty`
trick. This covers >80% of all uses of `st.write`.

A bit more context here:
https://github.com/streamlit/streamlit/pull/9562#discussion_r2023085071

## Testing Plan

- Added unit test. Existing tests should also cover this. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.