[theming] use markdown text colors for primitives (#11033)
## Describe your changes

With guidance from design (and from the discussion
[here](https://www.notion.so/snowflake-corp/Revamp-markdown-inline-code-colors-18a7170bb4168075966fc4900128af1e)),
the PR consolidated the primitive color codes (red/green/blue/etc...)
used in our themes with the Markdown text's color so that they are
consistent across the app.

Specifically, I replaced red/blue/green/yellow in the base (light) and
dark themes with the definition
[here](https://github.com/streamlit/streamlit/blob/develop/frontend/lib/src/theme/getColors.ts#L123-L127).
RFC if we need change the gray color as well. I left it out due to it
having multiple variants already.


## GitHub Issue Link (if applicable)

## Testing Plan

- E2E Tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.