Add devcontainer configuration (#10975)
## Describe your changes

Add an initial version of devcontainer. There are a couple of potential
follow-ups:
- Use uv for faster installation and find other ways to speed up the
initial installation.
- Tweak the Dockerfile to be closer to our CI runner (for debugging e2e
tests, maybe even generating snapshots -> related to
https://github.com/streamlit/streamlit/pull/10774)
- Optimize the VS-code settings. 
- Open two terminals, one with the frontend dev server and another one
with `streamlit hello` or some other app?
- Add a contribution readme that explains how one can start an app + dev
server and open it on startup.

Closes https://github.com/streamlit/streamlit/issues/7242


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

---------

Co-authored-by: Lukas Masuch <103002573+sfc-gh-lmasuch@users.noreply.github.com>