Reduce flakiness of st_radio_test by adding an expect before clicking. (#11129)
## Describe your changes

I have noticed that this test is flaky lately. I looked at the Radio.tsx
component and couldn't find any evidence that there is
unmounting/mounting or similar that would be causing the flakiness so
this PR adds an expect to harden the test.

## GitHub Issue Link (if applicable)

## Testing Plan

Test Only. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.