[Refactor] [AdvancedLayouts] Simplify width styling on VerticalBlock (#11081)
## Describe your changes

The purpose of this PR is to simplify some of the styling methods we
have in place to make it easier to roll out the extended styling options
that we are adding as part of Advanced Layouts.

**Removing `resizeObserver` from `VerticalBlock`.** 

Based on my investigation, I believe we needed this code primarily
because we used to prop-drill the container width to individual
elements. A previous PR removed this prop-drilling. Most elements now
use `width=100%` to expand to their container width. Some elements need
an explicit pixel width, but they now get this via `useCalculatedWidth`.

Containers (`ElementContainer` and `VerticalBlock`) will now also be set
to `width=100%` when they should stretch to fill the container
(`use_container_width=True`). If the user provides an explicit width,
they will be set to that. `maxWidth=100%` is added to prevent container
overflow. Image lists are a special case and the element container for
that must always be `width=100%`.

**Consolidating `VerticalBlockBorderWrapper` and
`StyledVerticalBlockWrapper`.**

Now that we are not observing the width here, we can consolidate these
two wrappers. A wrapper is still required for the scrolling behaviour
which requires `display: block`.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.