st.context.url improvement for multipage app (#11107)
## Describe your changes

Previously, the behaviour of `st.context.url` exposed whatever in the
address bar, but in a multipage app, the message could be sent with an
old URL (e.g., when clicking on navigation `Page 2` from `Page 3`).

This change does not expose the URL without processing, but rather
processes it a little bit by trimming the existing page suffix and
adding it based on the current page. That leads to correct behavior in
case of multipage apps too.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python) DONE!
- E2E Tests DONE!
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.