Support emoji favicons through navigation (#11154)
## Describe your changes

In #10912, I made a change that ensures emojis are verified from the
backend and informs the frontend through the `emoji:` prefix. This
inadvertently broke page icons in pages for `st.navigation`. This change
fixes that by detecting the emoji in the navigation call and prefix it
accordingly.

This fixes the favicon but is also used in the sidebar nav, so we needed
the `EmojiIcon` to support both emojis with and without the prefix.

## Testing Plan

- JS and Python unit tests updated
- E2E test should continue to show the same icons

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.