Fix content alignment for dataframe columns (#11152)
## Describe your changes

Fixes the unofficial support for configuring the content alignment of
columns via `st.column_config`. There was an issue that we didn't use
the correct property name for setting the content alignment for some
columns. This is fixed by enforcing correct type checking.

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/11110

## Testing Plan

- This isn't an official feature yet -> so we don't really have tests.
But the changes to the type checking should ensure that we are setting
correct properties.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.