Move min constraints to script folder to allow automated dependency updates (#11191)
## Describe your changes

Moves the min-constraint file to the scripts/assets folder so that the
automated dependency updates via dependabot in `/lib` work correctly.
The reason this is necessary is that there isn't any way in dependabot
config to ignore specific requirement files. And the min-constraints
isn't supposed to be updated.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.