Enable python integration tests for all PRs (#11200)
## Describe your changes

Since our CI secrets are not available in PRs from forks or dependabot,
we excluded the integration test job from running in these setups.
However, only a single integration test actually requires secrets.
Therefore, we are just skipping this single test only if the secrets are
empty.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.