Update mypy version and fix typing issues (#11295)
## Describe your changes

We were stuck on mypy 1.10 for a long time because of some typing
issues. This PR removes the upper pin for mypy and fixes all typing
issues with the newer version. I also added additional type checks to
extend coverage.

Closes https://github.com/streamlit/streamlit/pull/11239

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.